Add the change sketch board grid size feature
I added a button that changes the size of the grids in the sketch board.
The change grid size button will ask for a grid number per side as input.
Then, the existing grids will be removed from the sketch board.
The new grids will be generated based on the user input.